requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.51
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
# mysqlclient==2.2.0  # Windows 上推荐使用 pymysql
pymysql==1.1.0
netmiko==4.1.0
paramiko==3.4.0
httpx==0.25.2
python-multipart==0.0.5
# Excel处理依赖
pandas>=2.2.0
openpyxl==3.1.2
# Git操作依赖
gitpython==3.1.43
# CORS已在FastAPI内置，不需要单独安装

# 开发依赖
pytest==7.4.3
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1